% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_read_clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{eq_clean_data}
\usage{
eq_clean_data(filepath = NA, df = NA)
}
\arguments{
\item{filepath}{file path and name of file to load}

\item{df}{loaded data to work with}
}
\value{
tibble of earthquake data
}
\description{
load and clean earthquake data
}
\details{
This function takes either a file path with file name or already loaded data frame object. If a file path
is provided, the existence of a file at the specified path is checked, then the file is loaded and cleaned. If an
already loaded data frame object is provided, the data is cleaned. Data cleaning includes creating a date column,
noting if year is  BC or AD, and making sure latitude and longitude are numeric columns.
}
\examples{
\dontrun{
#Loading file
quake_data <- eq_clean_data(filepath = "inst/extdata/earthquakes.tsv")

#Working with loaded data
quake_data <- eq_clean_data(df = earthquakes)
head(quake_date)
}
}
