% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{eq_create_label}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{cleaned earthquake data}
}
\value{
modified data frame
}
\description{
Create Leaflet Popup with Location, Magnitude, and Total Deaths
}
\details{
This function creates a base option for more interesting leaflef map popups for cleaned significant
earthquake data. Popups created using this function will include location name(s) where earthquake occurred,
the magnitude of the earthquake, and the total deaths attributed to the earthquake.
}
\examples{
\dontrun{
eq_clean_data(filepath = "inst/extdata/earthquakes.tsv") \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(Country == "Mexico" & Date >= 2000) \%>\%
  dplyr::mutate(popup = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup")
}
}
