% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_read_clean_data.R
\name{eq_filtering}
\alias{eq_filtering}
\title{eq_filtering}
\usage{
eq_filtering(
  df,
  MinDate = NULL,
  MaxDate = NULL,
  SelectedCountry = NULL,
  groupingBy = NULL
)
}
\arguments{
\item{df}{data frame containing earthquake data}

\item{MinDate}{optional minimum date string of data to include}

\item{MaxDate}{optional maximum date string of data to include}

\item{SelectedCountry}{optional name of country to filter data by}

\item{groupingBy}{optional list of other data columns to include}
}
\value{
tibble of filtered earthquake data
}
\description{
Filter Earthquake Data to Desired Country and Data Columns
}
\details{
This function takes a cleaned earthquake data set and filters data to optional date ranges and country
of occurrence. An optional 'groupingBy' parameter allows for the selection of specified data columns in output.
}
\examples{
\dontrun{
# Selection of earthquakes within Argentina
Argentina_data <- eq_filtering(quake_data, SelectedCountry = "Argentina")

# Selection of earthquakes within Argentina since 1930
Argentina_data <- eq_filtering(quake_data, SelectedCountry = "Argentina", MinDate = "1930-01-01")

# Selection of earthquakes within Argentina with magnitude
Argentina_data <- eq_filtering(quake_data, SelectedCountry = "Argentina", groupingBy = c("Magnitude"))
}
}
