% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_read_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{data frame containing earthquake data}
}
\value{
tibble of earthquake data
}
\description{
set Location as country
}
\details{
This function takes an earthquake data set, checks for the existence of a Location*Name columns, then
extracts the country from the location column. The country is saved as the Location*Name data.
}
\examples{
\dontrun{
quake_data <- eq_location_clean(eq_date)
head(quake_date)
}
}
