% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{eq_map}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{cleaned data to work with}

\item{annot_col}{specific column to include in map popup}
}
\value{
leaflet map of earthquake locations
}
\description{
Create Leaflet Map of Earthquake Data
}
\details{
This function creates a leaflef map of cleaned significant earthquake data from  NOAA's significant
eathquake database (url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}). Data should be cleaned
using eq_clean_data() and eq_location_clean() for best usage. An annotation column can be specified for including
in a popup for the earthquake points.
}
\examples{
\dontrun{
eq_clean_data(filepath = "inst/extdata/earthquakes.tsv") \%>\%
  eq_location_clean() \%>\%
  dplyr::filter(Country == "Mexico" & Date >= 2000) \%>\%
  dplyr::mutate(popup = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup")
}
}
