% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{geom_timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetics mapping}

\item{data}{data to be used}

\item{stat}{stat objected with default "identity"}

\item{position}{position object with default "identity"}

\item{na.rm}{handling of NA values, removal default to FALSE}

\item{show.legend}{option to show or hide the legend with default set to FALSE}

\item{inherit.aes}{option to inherit aesthetics specified}

\item{...}{additional parameters passed to the function call such as color, alpha, or size}
}
\value{
a layer containing a `Geom*` object responsible for rendering the point-style graphic
}
\description{
geom_timeline plots a timeline of significant earthquakes
}
\details{
The graphic is a point-style plot showing significant earthquakes data originating from the NOAA database
(url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}). Earthquake data should be cleaned, and may be
filtered, to a specified date range or country. Earthquakes can be stratified by a factor variable, and points can
be formatted by color, point size, or alpha level based on user-specified data columns.
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}} #Date variable
  \item \code{y}          #Factor indicating some stratification
  \item \code{color}      #Color of points
  \item \code{size}       #Size of points
  \item \code{alpha}      #Transparency (1: opaque; 0: transparent)
}
}

\examples{
\dontrun{
#Basic plot:
quake_clean \%>\%
  ggplot() +
  geom_timeline(aes(x = Date)))

#Australian earthquakes colored by magnitude:
quake_clean \%>\%
 filter(Country == "Australia") \%>\%
 ggplot() +
 geom_timeline(aes(x = Date, color = Magnitude))
}
}
