% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_labels.R
\name{geom_timeline_labels}
\alias{geom_timeline_labels}
\title{geom_timeline_labels}
\usage{
geom_timeline_labels(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{aesthetics mapping}

\item{data}{data to be used}

\item{stat}{stat objected with default "identity"}

\item{position}{position object with default "identity"}

\item{na.rm}{handling of NA values, removal default to FALSE}

\item{show.legend}{option to show or hide the legend with default set to FALSE}

\item{inherit.aes}{option to inherit aesthetics specified}

\item{...}{additional parameters passed to the function call such as color, alpha, or size}
}
\value{
a layer containing a `Geom*` object responsible for rendering the point-style graphic
}
\description{
geom_timeline_labels adds a label to a timeline of significant earthquakes
}
\details{
The graphic is a text-style plot adding labels to points of a significant earthquakes timeline plot using
data orginating from the NOAA database (\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}.) Labels
require for size = Magnitude to be specified.
}
\section{Aesthetics}{

\code{geom_timeline} understands the following aesthetics (required aesthetics are in bold):
\itemize{
  \item \strong{\code{x}}     #Date variable
  \item \strong{\code{label}} #Factor for adding text labels
  \item \code{y}              #Factor indicating some stratification
  \item \code{n_max}          #Number of earthquake labels to show, ranked by magnitude
  \item \code{y_length}       #vertical line length to each data point
}
}

\examples{
\dontrun{
#Basic plot:
quake_clean \%>\%
  ggplot() +
  geom_timeline(aes(x = Date)))

#Australian earthquakes colored by magnitude:
quake_clean \%>\%
 filter(Country == "Australia") \%>\%
 ggplot() +
 geom_timeline(aes(x = Date, color = Magnitude))+
 geom_timeline_labels(aes(x = Date, label = Locale, size = Magnitude), n_max = 10)
}
}
