% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\arguments{
\item{GeomTimeline}{`Geom*` name created to plot earthquake data}
}
\value{
point-style graphic showing a timeline of earthqaukes
}
\description{
Creates a new `Geom*` to show timeline of significant earthquakes
}
\details{
The `Geom*` is created by using a cleaned data frame of earthquake data and plotting the
}
\keyword{datasets}
