% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_labels.R
\docType{data}
\name{GeomTimelineLabels}
\alias{GeomTimelineLabels}
\title{GeomTimelineLabels}
\format{
An object of class \code{GeomTimelineLabels} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimelineLabels
}
\arguments{
\item{GeomTimelineLabels}{`Geom*` name created to add lables to earthquake data}
}
\value{
text-style graphic with labels
}
\description{
GeomTimelineLabels
}
\keyword{datasets}
