% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_timeline.R
\name{pretty_timeline}
\alias{pretty_timeline}
\title{pretty_timeline}
\usage{
pretty_timeline(df, plot, timeline_y = FALSE)
}
\arguments{
\item{df}{cleaned data frame of earthquake data}

\item{plot}{existing ggplot object to apply formatting to}

\item{timeline_y}{if data was split on the y-axis by a factor column (TRUE/FALSE)}
}
\value{
modified timeline plot of earthquake data
}
\description{
Modify Existing Timeline Plot Formatting
}
\details{
This function modifes an existing timeline plot created using geom_timeline() of significant earthquake
data from NOAA's significant eathquake database (url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}).
The base plot should already exist. This function allows for consistent plot formatting.
}
\examples{
\dontrun{
Argentina_data <- eq_filtering(quake_clean, SelectedCountry = "Argentina", groupingBy = c("Deaths", "Magnitude"))
Arg_plot <- Argentina_data \%>\%
ggplot() +
geom_timeline(aes(x = Date, color = Magnitude, y = Deaths))
pretty_timeline(df = Argentina_data, plot = Arg_plot, timeline_y = TRUE)
}
}
