% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake_read_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{eq_location_clean}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{data frame containing earthquake data}
}
\value{
tibble of earthquake data
}
\description{
Addition of Country to Earthquake Data
}
\details{
This function takes an earthquake data set, checks for the existence of a Location*Name columns, then
extracts the country from the location column into a new 'Country' column. Additionally, a 'Locale' column is
created for within country location specification if applicable.
Special case handling currently includes: United Kingdom, United States
}
\examples{
\dontrun{
quake_data <- eq_location_clean(eq_date)
head(quake_data)
}
}
